Spike for switching to PVCs for scheduling
Currently the Stackable StickyScheduler is kept around to minimize the changes
on the ZK operator, but effectively ignored (nodeName is left unset, so that
K8s is left to schedule the pod itself).

stackable-pvc-operator should probably be moved into its own thing (or
integrated into agent?) eventually, but keeping it in here for now to keep the
spike centralized.